Add files via upload
Imagine you are in 1977 and you work as an air crash investigation officer on the Spanish island of Tenerife. You receive some bad news that two Boeing 747 passenger jets, KLM Flight 4805 and Pan Am Flight 1736, collided on the runway in the airport, which has resulted in the death of 583 people.
 
You have been tasked with investigating this incident to find the possible root causes for the crash.